+++ src/main/java/es/in2/wallet/domain/utils/ApplicationConstants.java
@@ -46,4 +46,7 @@ public class ApplicationConstants {
     public static final String VERIFIABLE_PRESENTATION_EXPIRATION_UNIT = "SECONDS";
     public static final String AUTH_SERVER_JWT_DECODER_PATH = "/protocol/openid-connect/certs";
     public static final String VAULT_HASHICORP_PATH = "/kv";
+    public static final String ENDPOINT_PIN = "/api/v1/pin";
+    public static final String ENDPOINT_HEALTH = "/health";
+    public static final String ENDPOINT_PROMETHEUS = "/prometheus";
 }