+++ src/main/java/es/in2/wallet/infrastructure/core/config/InternalCORSConfig.java
@@ -1,0 +1,33 @@
+package es.in2.wallet.infrastructure.core.config;
+
+import es.in2.wallet.application.ports.AppConfig;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.cors.CorsConfiguration;
+import org.springframework.web.cors.reactive.UrlBasedCorsConfigurationSource;
+
+import java.util.List;
+
+@Configuration
+@RequiredArgsConstructor
+public class InternalCORSConfig {
+
+    private final AppConfig appConfig;
+
+    @Bean("internalCorsSource")
+    public UrlBasedCorsConfigurationSource internalCorsConfigurationSource() {
+        CorsConfiguration configuration = new CorsConfiguration();
+
+        configuration.setAllowedOrigins(appConfig.getCorsAllowedOrigins());
+        configuration.setAllowedMethods(List.of("GET", "POST", "PUT", "DELETE", "OPTIONS"));
+        configuration.setAllowedHeaders(List.of("*"));
+        configuration.setAllowCredentials(true);
+        configuration.setMaxAge(3600L);
+
+        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+
+        source.registerCorsConfiguration("/**", configuration);
+        return source;
+    }
+}