+++ src/main/java/es/in2/wallet/infrastructure/core/config/WebSecurityConfig.java
@@ -3,8 +3,10 @@ package es.in2.wallet.infrastructure.core.config;
 import es.in2.wallet.application.ports.AppConfig;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.core.annotation.Order;
 import org.springframework.http.HttpMethod;
 import org.springframework.security.config.annotation.web.reactive.EnableWebFluxSecurity;
 import org.springframework.security.config.web.server.ServerHttpSecurity;
@@ -13,16 +15,21 @@ import org.springframework.security.oauth2.jwt.JwtValidators;
 import org.springframework.security.oauth2.jwt.NimbusReactiveJwtDecoder;
 import org.springframework.security.oauth2.jwt.ReactiveJwtDecoder;
 import org.springframework.security.web.server.SecurityWebFilterChain;
+import org.springframework.security.web.server.util.matcher.ServerWebExchangeMatchers;
+import org.springframework.web.cors.reactive.UrlBasedCorsConfigurationSource;
 
-import static es.in2.wallet.domain.utils.ApplicationConstants.GLOBAL_ENDPOINTS_API;
-import static org.springframework.security.config.Customizer.withDefaults;
+import static es.in2.wallet.domain.utils.ApplicationConstants.*;
 
+@Slf4j
 @Configuration
 @EnableWebFluxSecurity
 @RequiredArgsConstructor
-@Slf4j
 public class WebSecurityConfig {
+
     private final AppConfig appConfig;
+    private final PublicCORSConfig publicCORSConfig;
+    private final InternalCORSConfig internalCORSConfig;
+
     @Bean
     public ReactiveJwtDecoder jwtDecoder() {
         NimbusReactiveJwtDecoder jwtDecoder = NimbusReactiveJwtDecoder
@@ -30,25 +37,61 @@ public class WebSecurityConfig {
                 .jwsAlgorithm(SignatureAlgorithm.RS256)
                 .build();
         jwtDecoder.setJwtValidator(JwtValidators.createDefaultWithIssuer(appConfig.getAuthServerExternalUrl()));
-        log.debug(appConfig.getJwtDecoder());
-        log.debug(appConfig.getAuthServerExternalUrl());
+        log.debug("JWT Decoder URI: {}", appConfig.getJwtDecoder());
+        log.debug("JWT Issuer: {}", appConfig.getAuthServerExternalUrl());
         return jwtDecoder;
     }
+
+    // Filter chain for public endpoints
     @Bean
-    public SecurityWebFilterChain springSecurityFilterChain(ServerHttpSecurity http) {
+    @Order(1)
+    public SecurityWebFilterChain publicFilterChain(ServerHttpSecurity http) {
         http
+                .securityMatcher(ServerWebExchangeMatchers.matchers(
+                        ServerWebExchangeMatchers.pathMatchers(
+                                HttpMethod.GET,
+                                ENDPOINT_PIN,
+                                ENDPOINT_HEALTH,
+                                ENDPOINT_PROMETHEUS
+                        )))
                 .authorizeExchange(exchanges -> exchanges
-                                .pathMatchers("api/v1/pin").permitAll()
-                                .pathMatchers(HttpMethod.GET, GLOBAL_ENDPOINTS_API).authenticated()
-                                .pathMatchers(HttpMethod.POST, GLOBAL_ENDPOINTS_API).authenticated()
-                                .pathMatchers(HttpMethod.DELETE, GLOBAL_ENDPOINTS_API).authenticated()
-                                .anyExchange().permitAll()
-                ).csrf(ServerHttpSecurity.CsrfSpec::disable)
-                // Disables Cross-Site Request Forgery (CSRF) protection.
-                .httpBasic(ServerHttpSecurity.HttpBasicSpec::disable)
+                        .pathMatchers(
+                                HttpMethod.GET,
+                                ENDPOINT_PIN,
+                                ENDPOINT_HEALTH,
+                                ENDPOINT_PROMETHEUS
+                        ).permitAll()
+                        .anyExchange().authenticated()
+                )
+                .cors(cors -> cors.configurationSource(publicCORSConfig.publicCorsConfigSource()))
+                .csrf(ServerHttpSecurity.CsrfSpec::disable);
+
+
+        return http.build();
+    }
+
+
+
+    // Filter chain used by default, requires authentication
+    @Bean
+    @Order(2)
+    public SecurityWebFilterChain internalFilterChain(ServerHttpSecurity http) {
+
+        ReactiveJwtDecoder decoder = jwtDecoder();
+
+        http
+                .securityMatcher(ServerWebExchangeMatchers.pathMatchers(GLOBAL_ENDPOINTS_API))
+                .cors(cors -> cors.configurationSource(internalCORSConfig.internalCorsConfigurationSource()))
+                .authorizeExchange(exchanges -> exchanges
+                        .anyExchange().authenticated()
+                )
+                .csrf(ServerHttpSecurity.CsrfSpec::disable)
                 .oauth2ResourceServer(oauth2ResourceServer ->
                         oauth2ResourceServer
-                                .jwt(withDefaults()));
+                                .jwt(jwtSpec -> jwtSpec
+                                        .jwtDecoder(decoder))
+                );
+
         return http.build();
     }
 }