+++ CHANGELOG.md
@@ -3,11 +3,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-## [v1.4.4](https://github.com/in2workspace/in2-wallet-api/releases/tag/v1.4.4)
+## [v1.4.5](https://github.com/in2workspace/in2-wallet-api/releases/tag/v1.4.4)
 ### Changed
 - Implemented support for dcql_query and removed all remaining presentation_definition logic.
 - Added validations for credential format and id.
 - Introduced new custom exceptions for OID4VP.
+
+## [v1.4.4](https://github.com/in2workspace/in2-wallet-api/releases/tag/v1.4.3)
+### Changed
+- Add CORS configuration to health, prometheus and pin endpoints.
+- Remove Spring Webflux basepath parameter.
 
 ## [v1.4.3](https://github.com/in2workspace/in2-wallet-api/releases/tag/v1.4.3)
 ### Changed